site: adjust code structure
