+++ src/app/home/context-menu/context-menu.component.scss
@@ -1,0 +1,39 @@
+@import 'src/variables';
+@import '~le5le-ui/variables';
+
+.context-menus {
+  position: fixed;
+  z-index: 100;
+  color: #000;
+  background-color: #fff;
+  box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.3);
+  min-width: 1.5rem;
+  text-align: left;
+  padding: 0.08rem 0;
+
+  & > div {
+    line-height: 2.2;
+    a {
+      color: $color;
+      display: block;
+      padding: 0 0.2rem;
+      text-decoration: none;
+
+      &.flex {
+        display: flex;
+      }
+
+      &.disabled {
+        color: $color-gray !important;
+        cursor: default;
+      }
+    }
+  }
+
+  .line {
+    background-color: transparent !important;
+    padding: 0;
+    margin: 0.05rem 0;
+    border-top: 1px solid #eee;
+  }
+}
